Idlab

+++ app/Http/Controllers/PostController.php
@@ -95,7 +95,9 @@ class PostController extends Controller
     public function deletePost($postid)
     {
         $findpost = Post::find($postid);
-        $findpost->delete();
+        if ($findpost) {
+            $findpost->delete();
+        }
 
         return redirect(route('admin.home'))->with('success', 'Projeto atualizado com sucesso!');
     }
@@ -147,17 +149,11 @@ class PostController extends Controller
 
     //
     //Rotas de atividade
-    public function activityManagment()
-    {
-        $activities = activities::all();
-
-        return view('activitymanagment', compact('activities'));
-    }
-
     public function createActivity(Request $request)
     {
         $request->validate(
             [
+                'projname' => 'required|string| max: 100',
                 'projresume' => 'required|string|max:100',
                 'event_day' => 'required|date',
                 'image' => 'required|image|mimes:jpeg,png,jpg|max:2048'
@@ -167,6 +163,7 @@ class PostController extends Controller
         $imagePath = $request->file('image')->store('images', 'public');
 
         $activity = new activities([
+            'title' => $request->input('projname'),
             'resume' => $request->input('projresume'),
             'activity_date' => $request->input('event_day'),
             'image_path' => $imagePath
@@ -181,11 +178,81 @@ class PostController extends Controller
     {
         $findactivity = activities::find($postid);
         $findactivity->delete();
+        return redirect(route('admin.home'));
+    }
 
-        return response()->json(['message' => 'Atividade excluída com sucesso!'], 200);
+    public function editActivity($activityid)
+    {
+        $activity = activities::find($activityid);
+        return view('editactivity', compact('activity'));
+    }
+
+    public function putActivity(Request $request)
+    {
+        $request->validate(
+            [
+                'projname' => 'required|string| max: 100',
+                'projresume' => 'required|string|max:100',
+                'event_day' => 'required|date',
+                'image' => 'required|image|mimes:jpeg,png,jpg|max:2048'
+            ]
+        );
+
+        $imagePath = $request->file('image')->store('images', 'public');
+
+        $activity = new activities([
+            'title' => $request->input('projname'),
+            'resume' => $request->input('projresume'),
+            'activity_date' => $request->input('event_day'),
+            'image_path' => $imagePath
+        ]);
+
+        $activity->save();
+
+        return redirect('/activity/registeractivity')->with('success', 'Atividade editada com sucesso!');
     }
 
     //rotas da equipe
+    public function deletePartner($partnerid)
+    {
+        $findpartner = Team::find($partnerid);
+        if($findpartner)
+        {
+            $findpartner->delete();
+        }
+        return redirect(route('admin.home'));
+    }
+
+    public function editPartner($partnerid)
+    {
+        $partner = Team::find($partnerid);
+
+        return view('editpartner', compact('partner'));
+    }
+
+    public function putPartner(Request $request)
+    {
+        $request->validate([
+            'projname' => 'required|string|max:100',
+            'job' => 'required|string',
+            'projresume' =>'required|string|max: 100',
+            'image' => 'required|image|mimes:jpeg,png,jpg|max:2048',
+        ]);
+
+        $imagePath = $request->file('image')->store('images', 'public');
+
+        $team = new Team([
+            'title' => $request->input('projname'),
+            'resume' => $request->input('projresume'),
+            'job' => $request->input('job'),
+            'image_path' => $imagePath,
+        ]);
+
+        $team->save();
+
+        return redirect('/team/registerpartner')->with('success', 'Integrante editado com sucesso!');
+    }
+
     public function createPartner(Request $request)
     {
         $request->validate([
@@ -199,7 +266,7 @@ class PostController extends Controller
 
 
         $team = new Team([
-            'name' => $request->input('projname'),
+            'title' => $request->input('projname'),
             'resume' => $request->input('projresume'),
             'job' => $request->input('job'),
             'image_path' => $imagePath,
